Upgrade to `WinSparkle-0.8.0`
