+++ bin/generate_keys.dart
@@ -14,7 +14,7 @@ Future<void> main(List<String> arguments) async {
 
   String executable = Platform.isMacOS
       ? '${Directory.current.path}/macos/Pods/Sparkle/bin/generate_keys'
-      : '${Directory.current.path}\\windows\\flutter\\ephemeral\\.plugin_symlinks\\auto_updater\\windows\\WinSparkle-0.7.0\\bin\\generate_keys.bat';
+      : '${Directory.current.path}\\windows\\flutter\\ephemeral\\.plugin_symlinks\\auto_updater\\windows\\WinSparkle-0.8.0\\bin\\generate_keys.bat';
 
   Process process = await Process.start(
     executable,
