--- windows/WinSparkle-0.7.0/bin/sign_update.bat
@@ -1,11 +1,0 @@
-@echo off
-
-set argC=0
-for %%i in (%*) do set /A argC+=1
-
-if not "%argC%"=="2" (
-  echo Usage: %0 update_file private_key
-  exit /b 1
-)
-
-openssl dgst -sha1 -binary < "%~1" | openssl dgst -sha1 -sign "%~2" | openssl enc -base64
